requirements.txt
certifi==2023.11.17
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
polars==0.20.5
polygon-api-client==1.13.4
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
seaborn==0.13.1
six==1.16.0
tzdata==2023.4
urllib3==1.26.18
websockets==12.0
